backend/requirements.txt
Django==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
drf-spectacular==0.27.1
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0

